+++ src/ft_utils.c
@@ -6,18 +6,18 @@
 /*   By: hokutosuzuki <hosuzuki@student.42toky      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/07 07:14:49 by hokutosuz         #+#    #+#             */
-/*   Updated: 2022/02/14 15:17:17 by hokutosuz        ###   ########.fr       */
+/*   Updated: 2022/02/16 08:05:01 by hokutosuz        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 #include "../libft/libft.h"
 
-int	ft_intmax(t_stock *lst, unsigned long long total, int len)
+int	ft_intmax(t_stock *lst, size_t total, size_t len)
 {
 	if (len <= 0)
 		return (GOOD);
-	if ((unsigned long long)INT_MAX <= total + (unsigned long long)len)
+	if ((size_t)INT_MAX <= total + len)
 	{
 		lst->status = ERROR;
 		return (ERROR);
@@ -25,7 +25,7 @@ int	ft_intmax(t_stock *lst, unsigned long long total, int len)
 	return (GOOD);
 }
 
-int	ft_write(t_stock *lst, char	*content, int len)
+int	ft_write(t_stock *lst, char	*content, size_t len)
 {
 	if (ERROR == ft_intmax(lst, lst->total_len, len))
 		return (ERROR);
