+++ src/ft_width_precision.c
@@ -6,7 +6,7 @@
 /*   By: hokutosuzuki <hosuzuki@student.42toky      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/02 20:41:14 by hokutosuz         #+#    #+#             */
-/*   Updated: 2022/02/14 15:59:30 by hokutosuz        ###   ########.fr       */
+/*   Updated: 2022/02/16 07:49:58 by hokutosuz        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,14 +15,14 @@
 
 void	ft_precision(const char *fmt, t_stock *lst, size_t *i)
 {
-	size_t	len;
+	long long	len;
 
 	if (fmt[*i] != '.')
 		return ;
 	(*i)++;
 	if (fmt[*i] == '*')
 	{
-		lst->precision = va_arg(lst->args, int);
+		lst->precision = va_arg(lst->args, long long);
 		if (ERROR == ft_intmax(lst, lst->total_len, lst->precision))
 			lst->status = ERROR;
 		(*i)++;
@@ -42,11 +42,11 @@ void	ft_precision(const char *fmt, t_stock *lst, size_t *i)
 
 void	ft_width(const char *fmt, t_stock *lst, size_t *i)
 {
-	size_t	len;
+	long long	len;
 
 	if (fmt[*i] == '*')
 	{
-		lst->width = va_arg(lst->args, int);
+		lst->width = va_arg(lst->args, long long);
 		if (lst->width < 0)
 		{
 			lst->width = -(lst->width);
